pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioconsole>=0.8.2",
    "colorama>=0.4.6",
    "langgraph>=0.3.34",
    "openai>=2.29.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.7",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "ASYNC"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
